refactor: LoginMember 어노테이션 위치 변경
