+++ src/main/java/auth/resolver/LoginMember.java
@@ -1,4 +1,4 @@
-package auth.login;
+package auth.resolver;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
