+++ src/main/java/nextstep/config/WebMvcConfiguration.java
@@ -4,6 +4,7 @@ import auth.interceptor.AdminInterceptor;
 import auth.util.JwtTokenProvider;
 import auth.resolver.LoginMemberArgumentResolver;
 import lombok.AllArgsConstructor;
+import nextstep.web.member.domain.Member;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
@@ -16,7 +17,7 @@ import java.util.List;
 public class WebMvcConfiguration implements WebMvcConfigurer {
 
     private final JwtTokenProvider jwtTokenProvider;
-    private final LoginMemberArgumentResolver loginMemberArgumentResolver;
+    private final LoginMemberArgumentResolver<Member> loginMemberArgumentResolver;
 
     @Override
     public void addInterceptors(InterceptorRegistry registry) {
